some edits and back page at admin

+++ database/migrations/2019_11_23_143713_create_sub_categories_table.php
@@ -20,7 +20,7 @@ class CreateSubCategoriesTable extends Migration
             $table->string('slug_en')->index()->nullable();
             $table->string('slug_ar')->index()->nullable();
             $table->string('image')->nullable();
-            $table->unsignedTinyInteger('category_id')->nullable();
+            $table->unsignedInteger('category_id')->nullable();
             $table->unsignedTinyInteger('sort')->nullable();
             $table->boolean('status');
             $table->softDeletes();
